+++ Unity Utils/Assets/Testing Scripts/TestingScript2.cs
@@ -5,6 +5,7 @@ using UnityUtils.ScriptUtils.Objects;
 using UnityEngine.SceneManagement;
 using UnityUtils.ScriptUtils.Cameras;
 using JetBrains.Annotations;
+using UnityEngine.InputSystem;
 
 public class TestingScript2 : MonoBehaviour
 {
@@ -29,6 +30,16 @@ public class TestingScript2 : MonoBehaviour
         modifierTest.AddTemporaryModifier(new ObjectModifierData(ObjectModifiers.ModifierType.Flat, 3), 1);
 
         ObjectDelays.CallFunctionAfterTime(() => CameraShake.Screenshake(intensity: 5), 3);
+        ObjectDelays.CallFunctionAfterTime(() => CameraShake.Screenshake(intensity: 5), 3.2f);
+        ObjectDelays.CallFunctionAfterTime(() => CameraShake.Screenshake(intensity: 5), 4);
         ObjectDelays.CallFunctionAfterTime(() => GetComponent<ObjectColorFlash>().FlashColor(Color.blue, 2), 3);
+    }
+
+    private void Update()
+    {
+        if (Keyboard.current.hKey.wasPressedThisFrame)
+        {
+            CameraShake.Screenshake();
+        }
     }
 }